Move most of nft support to coin-framework
